implement $edit() for sc_json
add a method for editing the json file
and an update() method to the sc_table
class.

The json class now also has active bindings
which parse the components of the json
file (database, fields and dimensions)

all parsers (inclunding the print method)
reflect the state of the modified file.