sc_table: inherit from sc_data
restructure sc_table to make it compatible
with the base class sc_data. In the new
version, the main part of tabulate() is
executed in the base class which provides
more flexibility.

- mixtures of fields with totals and without
  totals are allowed
- parameters can be specified as codes or
  as labels
- parameter raw can be used to return
  codes instead of labels

The sc_table class now also inherits
total_codes() and a more flexible
implementation of field()

$data and $meta are now parsed eagerly
which means that the slots are calculated
at construction time rather than the first
access

there was a regression regarding the
annotations parameter. Annotations are
stored as attributes in $data_raw and
dropped during aggregation. They currently
don't have priority and will be
re-implemented at some point