special handling of totals: time variables
time variables now use the latest
observation as totals. This is done
by pushing the data to the indices of
the total lookup.

The current logic asumes that no totals
are present - if they are the will be
overwritten

In case of multiple time variables, this
logic might also break because the totals
are pushed in an iterative fashion rather
than using a max-filter for all columns
simultaneously

TODO:

- test with multiple time variables
- check if the totals are non-empty for
  certain cubes