update urs in sc_browse*() namespace
there is now a check to determine
wether STATcubeR is used inside the
firewall. If not, sc_browse() and
sc_browse_prefernces() now forwad to
the external STATcube and ensure that
the user is logged in when browsing
STATcube.

This commit also affects the $browse()
method of sc_table objects, which will
now determine the server from the json
request and open the correct deep-link
to the databases info page

possible TODOs:
- throw errors if server is not "ext"
  and is_stat() is FALSE
- use an environment variable in
  is_stat() rather than relying on
  Sys.info(). Otherwise, this the browse
  functions will only work on the two
  main R servers