update od_table article
add tooltips for csv and json resources

reorder the sections and put the `$tabulate()` method in earlier. This makes it easier to tak about the metadata

explain the print output of od_table objects. The print method was recently extended to show the number of classification elements

also add a tooltip in od_list.Rmd since this already showcases a resource file

more examples for the usage of od_json so it is easier to see how the CLI correspons to the file structure