add a cli theme for {pkgdown}
add a hook in .onLoad() which customizes the
cli theme in case pkgown is running.
Setting this here rather than in
vignettes/R/df_print.R has the advantage that
@examples are also affected by the styling

- options(cli.theme) is currently used only for
  colors
- hyperlink support is enabled even if the console
  doesn't support it because after the conversion
  into HTML everything is fine
- The same is true for the limit of 256 colors
- ansi warnings because of incompatible colors
  (and possibily links) are suppressed

TODO: Not sure if using .onLoad() this way is
considered good practice. A "cleaner" away would
be to define helper function which is capable
of setting (and unsetting) the styles