build external links with devtools::document()
assume that the environment where document()
runs can be characterized by the NOT_CRAN
flag being present (not necessarily FALSE or
TRUE). This is to avoid having links to the
editing server in the manpages and on pkgdown

this will only matter if the docs are built
inside the STAT firewall. If at some point,
the docs are built via gh-actions, this can be
reverted