add httptest to check error handling
there are now some additional tests which
were inspired by vignettes/sc_last_error.Rmd
and check wether the error-types are as
expected

those tests use httptest. The contents of
tests/testthat/statcube* are cached API
responses which will be used inside
the new tests.

the only modification of the caches prior to
versioning them was that the support id was
removed from some responses.