od_table_local(): support $PublDateTime$
some internal datasets now use
$PublDateTime$ as a placeholder for
the deployment timestamp. Make sure
that those datasets can be used with
STATcubeR

The way this is implemented now,
reading and resaving a dataset is not
a no-op because the interpolated
value will be written in place of the
placholder. There might come a point
where it makes sense to implement this
differently in order to preserve the
placeholder

[ci skip]